Added Next and Prev properties for next and prev in VisualNodes. Implemented traceback to find body of visualNode.
Traceback will make adding to else much easier. The setters and getters will allow me to override what next means depending on kind of visual node, for example on VisualBlocks, next is actually next of it's inner list, not  it's own next. Whether that's semantically a good choice is still in mind so kept nextChild and prevChild methods just incase of switching.